requirements.txt

flask==3.0.3
flask-restful==0.3.10
flask-limiter==3.8.0
langchain==0.3.1
langchain-community==0.3.1
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4
huggingface-hub==0.25.1
sentence-transformers==3.1.1
chromadb==0.5.5
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
typing-extensions==4.12.2